8384109: Add value class array cases to NegativeArraySizeExceptionTest
